Add tests for element_at, split, and aggregate functions
